moved ENABLE_NEGATIVE_TESTING into the Flags object
